refactor: extract and rename more
